Provide cleaned up small icons for DACAI screen
The stock icons of the DACAI firmware have strong JPG artifacts. For the small
icons this results in the background transparency "black" not being recognized
correctly in various degrees. In the worst case, not at all, so some icons had
black boxes around them.

This patch is a quick, albeit not perfect cleanup job reducing these artifacts on
the small icons and the CREALITY logo significantly, eliminating all black boxes
and most stray black pixels.